Solve day 2 part 2 exercise
